requirements.txt
asgiref==3.4.1
black==21.8b0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
cryptography==3.4.8
defusedxml==0.7.1
Django==3.2.7
django-allauth==0.45.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
flake8==3.9.2
gunicorn==20.1.0
idna==3.2
mccabe==0.6.1
mypy-extensions==0.4.3
oauthlib==3.1.1
pathspec==0.9.0
platformdirs==2.3.0
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyJWT==2.1.0
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
regex==2021.8.28
requests==2.26.0
requests-oauthlib==1.3.0
sqlparse==0.4.2
tenacity==8.0.1
tomli==1.2.1
typing-extensions==3.10.0.2
urllib3==1.26.6
coreapi==2.3.3
